💄 add kobo toolbox logo
