+++ apps/hpc-cdm/src/assets/logos/enketologo.png
